DisplayStrings are now always valid UTF-8
Metric labels and string values are expected to be always valid
UTF8 strings, which is not alwys the case for overridden types
or even buggy snmp implementations. So instead of dropping the
whole pull result, invalid bytes get replaced in DisplayStrings
by a hand sign (0xE29C8B alias <U+270B>) and thus can be displayed
and are valid UTF8.