new feature: fallback_label for module + metric
Per default the exporter generates and injects a label for
non-numeric metric values having the same name as the metric and
using the metric's value as label value.

With this option one may override the label name to use with
fallback_label: label_name  in the generator and exporter config.

It can be set at the module level as a catch all setting, or one
may specify it in the overrides section of the desired metric. The
letter takes precedence over the module setting.